+++ src/mailman/docs/NEWS.rst
@@ -107,6 +107,8 @@ Commands
 
 Bug fixes
 ---------
+ * Subscription disabled probe warning notification messages are now sent
+   without a `Precedence:` header.  Given by Mark Sapiro. (LP: #808821)
  * Fixed KeyError in retry runner, contributed by Stephen A. Goss.
    (LP: #872391)
  * Fixed bogus use of `bounce_processing` attribute (should have been